+++ test/python/evaluation/statistics/__init__.py
@@ -1,6 +1,6 @@
 from test.python.evaluation import CURRENT_TEST_DATA_FOLDER
 
-STATISTICS_TEST_DATA_FOLDER = CURRENT_TEST_DATA_FOLDER / 'statistics'
+STATISTICS_TEST_DATA_FOLDER = CURRENT_TEST_DATA_FOLDER / 'issues_statistics'
 
 GET_RAW_ISSUES_DATA_FOLDER = STATISTICS_TEST_DATA_FOLDER / 'get_raw_issues'
