+++ test/python/evaluation/statistics/test_get_raw_issues.py
@@ -6,7 +6,7 @@ from typing import List, Optional
 import pandas as pd
 import pytest
 from src.python.evaluation.common.pandas_util import get_solutions_df_by_file_path
-from src.python.evaluation.statistics.get_raw_issues import _filter_issues, _get_output_path, inspect_solutions
+from src.python.evaluation.issues_statistics.get_raw_issues import _filter_issues, _get_output_path, inspect_solutions
 from src.python.review.inspectors.inspector_type import InspectorType
 from src.python.review.inspectors.issue import BaseIssue, CodeIssue, IssueType, LineLenIssue, MaintainabilityLackIssue
